Complete Part 4

+++ part4/blog-list/controllers/blogs.js
@@ -1,18 +1,37 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog
+    .find({}).populate('user', { username: 1, name: 1, id: 1 })
   response.json(blogs)
 })
 
 blogsRouter.post('/', async (request, response) => {
-  const blog = new Blog(request.body)
-  const result = await blog.save()
-  response.status(201).json(result)
+  const body = request.body
+
+  const user = request.user
+
+  const blog = new Blog({
+    ...body,
+    likes: body.likes || 0,
+    user: user._id
+  })
+
+  const savedBlog = await blog.save()
+
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+
+  response.status(201).json(savedBlog)
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const blog = await Blog.findById(request.params.id)
+  if (blog.user.toString() !== request.user._id.toString()) {
+    return response.status(401).json({ error: 'blog can only be deleted by user who wrote it' })
+  }
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
